Create GenerateDataBase class and methods to save tracks data

+++ src/main/java/com/packt/app/genre/GenreRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource
-public interface GenreRepository extends CrudRepository<Genre, Integer> {
+public interface GenreRepository extends CrudRepository<Genre, Long> {
 
 }
