+++ src/main/java/com/packt/app/playlist/PlaylistRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource
-public interface PlaylistRepository extends CrudRepository<Playlist, Integer> {
+public interface PlaylistRepository extends CrudRepository<Playlist, Long> {
 
 }
